Refactoring + test folder README.md
