+++ tests/README.md
@@ -1,0 +1,13 @@
+# Tests folder
+
+To run tests, install the re queirements and run
+
+```bash
+~$ python -m unittest discover ./tests
+```
+
+To run specific types of tests, for example unit test run
+
+```bash
+~$ python -m unittest discover ./tests/unit
+```
